Have describe write its HTML to a temporary file.
This is more resilient than writing it into the *Racket REPL Redirected
Output* buffer, because comint-redirect-send-command-to-process looks
for comint-prompt-regexp to detect end of output. If a line of HTML
starts with something that looks like that -- e.g. ">" or "p>" -- then
it will break. Attempting to make comint-prompt-regexp resilient to that
is a losing game. Instead, it's easy to write the HTML to a file, and
print to the REPL output simply a string that is the file pathname.